# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/auth/verfiy":{"post":{"operationId":"AuthController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAuthDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/users":{"post":{"operationId":"UsersController_create","summary":"Create user","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"example":{"email":"john.smith@gmail.com","role":"patient","company_name":"test","status":true,"deleted_at":null,"id":0,"created_at":"2023-05-29T09:24:58.517Z","updated_at":"2023-05-29T09:24:58.517Z"}}}}},"403":{"description":"Forbidden."}},"tags":["User"],"security":[{"bearer":[]}]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["User"],"security":[{"bearer":[]}]}},"/api/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"],"security":[{"bearer":[]}]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["User"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"],"security":[{"bearer":[]}]}}},"info":{"title":"Groww Business","description":"Empowering entrepreneurs with customizable e-commerce.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateAuthDto":{"type":"object","properties":{"email":{"type":"string","example":"groww@businuss.com"},"phone_number":{"type":"string","example":"9876587432"},"user_type":{"type":"string","enum":["customer","admin","owner","super_owner"]},"company_name":{"type":"string","example":"groww business"},"status":{"type":"boolean","example":1},"otp":{"type":"string","example":"556655"},"token":{"type":"string","example":"$2b$20$cQ25ot5TqAIGn1Pic2rCNenN2aDrFAxk83OKlvioKFskwVwcJ5R8O"}},"required":["email","phone_number","user_type","company_name","status","otp","token"]},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"groww@businuss.com"},"phone_number":{"type":"string","example":"9876587432"},"user_type":{"type":"string","enum":["customer","admin","owner","super_owner"]},"company_name":{"type":"string","example":"groww business"},"status":{"type":"boolean","example":1}},"required":["email","phone_number","user_type","company_name","status"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string","example":"groww@businuss.com"},"phone_number":{"type":"string","example":"9876587432"},"user_type":{"type":"string","enum":["customer","admin","owner","super_owner"]},"company_name":{"type":"string","example":"groww business"},"status":{"type":"boolean","example":1}}}}}}